Finished convert feature (#28)
* Finished convert feature

* Update build-android.yaml

* Remove unsound null-safety from ci

+++ lib/service/services/swap/model/execute_quote/execute_quote_request_model.dart
@@ -7,11 +7,11 @@ part 'execute_quote_request_model.g.dart';
 class ExecuteQuoteRequestModel with _$ExecuteQuoteRequestModel {
   const factory ExecuteQuoteRequestModel({
     required String operationId,
-    required int price,
+    required double price,
     required String fromAsset,
     required String toAsset,
-    @JsonKey(name: 'fromAssetVolume') required int fromAssetAmount,
-    @JsonKey(name: 'toAssetVolume') required int toAssetAmount,
+    @JsonKey(name: 'fromAssetVolume') required double fromAssetAmount,
+    @JsonKey(name: 'toAssetVolume') required double toAssetAmount,
     required bool isFromFixed,
   }) = _ExecuteQuoteRequestModel;
